// packages/scs/meraki-poller/src/features/settings/types/form.ts
export type FieldValues = {
   Device: number;
   Interface: number;
   Wireless: number;
   Client: number;
   Application: number;
   Uplink: number;
   VPN: number;
   Topology: number;
   rate_limit: number;
   ['datatype-wireless_latency_stats']: boolean;
   ['datatype-wireless_connection_stats']: boolean;
   ['datatype-client_configuration']: boolean;
   ['datatype-client_applications']: boolean;
   ['datatype-switch_ports']: boolean;
   ['datatype-events']: boolean;
   ['datatype-topology']: boolean;
   config_poll_interval: number;
};